List below: 1. acceptOrder Controller/Service/DataValidation 2. assignRoom -||- 3. deleteEmployee -||- 4. Minor modifications in other files

+++ HotelManagement/src/main/java/com/hotelmangementprogram/hotelmanagement/model/Menu.java
@@ -17,5 +17,5 @@ public class Menu {
     @Column(name = "dishName")
     private String dishName;
     @Column(name = "dishPrice")
-    private double dishPrice;
+    private Float dishPrice;
 }
